Support buildpack builds
To support building the application via buildpacks the application configuration
was updated to read from environment variables. A procfile was added to describe
how to startup the application(s). The procfile depends on the `APP_SUBPROJECT`
environment variable to determine which app (authorization-server or
resource-server) to run.

The gradle configuration updated to support the `run` task by including the
"application" plugin and defining the main class of the applications.

The Dockerfiles were updated to add each of the required environment variables.
Variables were placed one-per-line to support Docker 1.3.